test: add BREAD integration tests for backend resources

+++ sw-be-container/src/models.rs
@@ -4,6 +4,7 @@ use sqlx::FromRow;
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, sqlx::Type)]
 #[sqlx(type_name = "user_role", rename_all = "lowercase")]
+#[serde(rename_all = "lowercase")]
 pub enum Role {
     User,
     Support,
@@ -126,6 +127,7 @@ pub struct FertiliserApplication {
     pub is_protected_urea: Option<bool>,
     pub buffer_zone_confirmed: Option<bool>,
     pub evidence_of_control: Option<String>,
+    #[sqlx(default)]
     pub geometry_wkt: Option<String>,
     pub updated_at: Option<DateTime<Utc>>,
     pub is_deleted: Option<bool>,
@@ -133,8 +135,8 @@ pub struct FertiliserApplication {
 
 #[derive(Serialize, Deserialize, Clone, Debug, FromRow)]
 pub struct OrganicManureApplication {
-    pub id: Option<i64>,
-    pub event_id: i64,
+    pub id: Option<i32>,
+    pub event_id: i32,
     pub manure_type: String,
     pub volume_applied_m3_per_ha: Option<f64>,
     pub weight_applied_tonnes_per_ha: Option<f64>,
@@ -144,6 +146,7 @@ pub struct OrganicManureApplication {
     pub buffer_zone_distance_meters: Option<i32>,
     pub equipment_used: Option<String>,
     pub lesse_exemption_reason: Option<String>,
+    #[sqlx(default)]
     pub geometry_wkt: Option<String>,
     pub updated_at: Option<DateTime<Utc>>,
     pub is_deleted: Option<bool>,
@@ -151,8 +154,8 @@ pub struct OrganicManureApplication {
 
 #[derive(Serialize, Deserialize, Clone, Debug, FromRow)]
 pub struct ComplianceBreach {
-    pub id: Option<i64>,
-    pub farm_id: i64,
+    pub id: Option<i32>,
+    pub farm_id: i32,
     pub breach_type: String,
     pub severity: String,
     pub estimated_penalty_percentage: Option<f64>,
